- Add x86_64 support to vm_x86.c - Fix warning on mingw64

+++ src/qcommon/vm_local.h
@@ -157,7 +157,7 @@ struct vm_s {
 	int			entryOfs;
 	int			codeLength;
 
-	int			*instructionPointers;
+	intptr_t	*instructionPointers;
 	int			instructionCount;
 
 	byte		*dataBase;
